Create custom hook for event handler logic for user profile image uploading
UserProfileInfo and UserProfileEdit contain the exact
same logic for uploading user profile images,
and setting the alert text/opening the modal. Therefore,
it made sense to me to pull this logic out into a hook for
use in those components.

I do have concerns about whether this is a good design
practice but yolo who's gonna read this anyway lol